requirements.txt
# =============================================================================
# Core AI
# =============================================================================
anthropic>=0.40.0           # Claude API + beta tool runner

# =============================================================================
# Snowflake
# =============================================================================
snowflake-connector-python[pandas]>=3.6.0
snowflake-sqlalchemy>=1.6.0

# =============================================================================
# Data processing
# =============================================================================
pandas>=2.1.0
pydantic>=2.5.0

# =============================================================================
# Notifications
# =============================================================================
requests>=2.31.0            # Slack + Teams webhooks

# =============================================================================
# Config
# =============================================================================
pyyaml>=6.0.1

# =============================================================================
# Airflow (only needed on the Airflow worker, not in local dev)
# Comment out if running locally without Airflow
# =============================================================================
# apache-airflow>=2.8.0

# =============================================================================
# Dev / testing
# =============================================================================
pytest>=7.4.0
pytest-mock>=3.12.0
python-dotenv>=1.0.0        # Load .env in local development
